backend/pyproject.toml
[project]
name = "cdcosp-dummy-bank-api"
version = "0.1.0"
description = "CDC OSP Dummy Bank API - A FastAPI-based mock banking system for testing payment integrations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.120.1",
    "fastapi-permissions>=0.2.7",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "minio>=7.2.18",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-decouple>=3.8",
    "python-jose[cryptography]>=3.3.0",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
